restructure web services. does not build because of idle service tests
