+++ package.json
@@ -8,14 +8,11 @@
     "dist"
   ],
   "scripts": {
-    "build": "cd src && npm run compile",
-    "link-build": "npm link && cd testing && npm link xynga-web-services",
-    "preversion": "npm run build",
-    "postversion": "git push",
-    "rebuild": "npm run build && npm start",
-    "setup": "cd src && npm i && cd .. && npm run build && cd testing && npm i && cd .. && npm run link-build",
-    "start": "cd testing && start chrome http://localhost:4200 && ng serve",
-    "test": "cd testing && ng test --watch=false --code-coverage"
+    "start": "",
+    "sandbox": "ng serve",
+    "build": "ngc -p \"./tsconfig.module.json\"",
+    "test": "ng test",
+    "e2e": "ng e2e"
   },
   "author": "",
   "license": "MIT",
@@ -34,5 +31,40 @@
     "@angular/router": "^4.0.0 || ^5.0.0",
     "@ng-idle/core": "^2.0.0-beta.15",
     "angular2-notifications": "0.9.7"
+  },
+  "devDependencies": {
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.2.11",
+    "@angular/http": "^5.2.11",
+    "@angular/platform-browser": "^5.2.11",
+    "@angular/platform-browser-dynamic": "^5.2.11",
+    "@angular/router": "^5.2.11",
+    "core-js": "^2.5.7",
+    "rxjs": "^5.5.11",
+    "zone.js": "^0.8.26",
+    "@angular/cli": "1.6.6",
+    "@angular/compiler-cli": "^5.2.11",
+    "@angular/language-service": "^5.2.11",
+    "@ng-idle/core": "^2.0.0-beta.15",
+    "@types/jasmine": "~2.5.53",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "^6.0.112",
+    "angular2-notifications": "0.9.7",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.4.2"
   }
 }
