+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import {RouterModule, Routes } from "@angular/router";
 
 import { AppComponent } from './app.component';
-import { ServiceModule } from 'xynga-web-services';
+import { ServiceModule } from '../web-services';
 
 const routes: Routes = [{ path: 'testing', component: AppComponent }];
