--- src/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "xynga-web-services-src",
-  "version": "0.0.1",
-  "description": "",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "compile": "ngc"
-  },
-  "author": "",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xynga/web-services.git"
-  },
-  "devDependencies": {
-    "@angular/animations": "^4.0.0 || ^5.0.0",
-    "@angular/common": "^4.0.0 || ^5.0.0",
-    "@angular/compiler": "^4.0.0 || ^5.0.0",
-    "@angular/compiler-cli": "^4.0.0 || ^5.0.0",
-    "@angular/core": "^4.0.0 || ^5.0.0",
-    "@angular/http": "^4.0.0 || ^5.0.0",
-    "@angular/platform-browser": "^4.0.0 || ^5.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0 || ^5.0.0",
-    "@angular/router": "^4.0.0 || ^5.0.0",
-    "@ng-idle/core": "^2.0.0-beta.15",
-    "angular2-notifications": "0.9.7",
-    "rxjs": "^5.5.11",
-    "typescript": "^2.9.1",
-    "zone.js": "^0.8.26"
-  }
-}
